examples:testing box2d integration
