--- Examples/imgui.ini
@@ -1,21 +1,0 @@
-[Window][Debug##Default]
-Pos=60,60
-Size=400,400
-Collapsed=0
-
-[Window][Dear ImGui Demo]
-Pos=0,19
-Size=286,581
-Collapsed=0
-DockId=0x00000001,0
-
-[Window][DockSpace Demo]
-Pos=0,0
-Size=800,600
-Collapsed=0
-
-[Docking][Data]
-DockSpace   ID=0x3BC79352 Window=0x4647B76E Pos=0,19 Size=800,581 Split=X
-  DockNode  ID=0x00000001 Parent=0x3BC79352 SizeRef=286,581 Selected=0xE927CF2F
-  DockNode  ID=0x00000002 Parent=0x3BC79352 SizeRef=512,581 CentralNode=1
-
